feat: adds babel transform func to webpack loader
