+++ packages/webpack-loader/src/index.tsx
@@ -1,11 +1,19 @@
 import path from 'path';
+import { transform } from '@compiled/babel-plugin';
 
-export default function compiledLoader(this: any, content: string): string {
-  console.log('BUNDLING', content);
+/**
+ * Compiled webpack loader.
+ *
+ * @param this
+ * @param content
+ */
+export default async function compiledLoader(this: any, content: string): Promise<void> {
+  const callback = this.async();
+  const result = await transform(content);
 
-  this.addDependency(
-    path.normalize('/Users/mdougall/projects/compiled/examples/packages/webpack/src/app.js')
-  );
+  result.includedFiles.forEach((file) => {
+    this.addDependency(path.normalize(file));
+  });
 
-  return content;
+  callback(null, result.code);
 }
